Revert "Move SiteConfig helper functions into dedicated file"
This reverts commit cf853fbc4ae233d331963a4b1b2cb3f1fe6b4adf.
