+++ site/src/app/[domain]/sitemap.ts
@@ -1,6 +1,6 @@
 import { gql } from "@comet/cms-site";
+import { getSiteConfig } from "@src/middleware";
 import { createGraphQLFetch } from "@src/util/graphQLClient";
-import { getSiteConfig } from "@src/util/SiteConfig";
 import { MetadataRoute } from "next";
 
 import { GQLPrebuildPageDataListSitemapQuery, GQLPrebuildPageDataListSitemapQueryVariables } from "./sitemap.generated";
